# api/apigateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Album and image management apis",
        "title": "Album Management",
        "contact": {
            "name": "API Support",
            "email": "ypankaj007@gmail.com"
        },
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/albums/{albumId}": {
            "post": {
                "description": "Create album to store the images",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Create album",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Album data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.albumReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Get album by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Get album",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        },
        "/albums/{albumId}/get": {
            "get": {
                "description": "Get album by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "Get album",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.albumResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        },
        "/albums/{albumId}/images": {
            "get": {
                "description": "List image in album",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "List image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.listImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Create Image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "images",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        },
        "/albums/{albumId}/images/{imageId}": {
            "delete": {
                "description": "Get image by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Get image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Image ID",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        },
        "/albums/{albumId}/images/{imageId}/get": {
            "get": {
                "description": "Get image by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Get image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Image ID",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.imageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        },
        "/albums/{albumId}/list": {
            "get": {
                "description": "List album in album",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "album"
                ],
                "summary": "List album",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent Album ID",
                        "name": "albumId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.listAlbumResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.basicResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.albumReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My Collection"
                }
            }
        },
        "handler.albumResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "ParentAlbumId": {
                            "type": "integer"
                        },
                        "createdAt": {
                            "type": "integer"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "path": {
                            "type": "string"
                        },
                        "updatedAt": {
                            "type": "integer"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.basicResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.imageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "albumId": {
                            "type": "integer"
                        },
                        "createdAt": {
                            "type": "integer"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "updatedAt": {
                            "type": "integer"
                        },
                        "url": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.listAlbumResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "count": {
                            "type": "integer"
                        },
                        "list": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "ParentAlbumId": {
                                        "type": "integer"
                                    },
                                    "createdAt": {
                                        "type": "integer"
                                    },
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "path": {
                                        "type": "string"
                                    },
                                    "updatedAt": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.listImageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "count": {
                            "type": "integer"
                        },
                        "list": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "albumId": {
                                        "type": "integer"
                                    },
                                    "createdAt": {
                                        "type": "integer"
                                    },
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "updatedAt": {
                                        "type": "integer"
                                    },
                                    "url": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}